Added most of the docstring for controlPhaseSLM.py, reoganisation of the structure
