+++ santec_driver/__init__.py
@@ -2,7 +2,6 @@
 # -*- coding:utf-8 -*-
 # -*-mode:python ; tab-width:4 -*- ex:set tabstop=4 shiftwidth=4 expandtab: -*-
 
-
 __all__ = ["gxwrapper", "dxwrapper", "gxiapi", "gxidef"]
 
 __version__ = '2.0.2105.9261'
