--- settings.py
@@ -1,7 +1,0 @@
-wavelength = 800e-9
-
-slm_size = (1200, 1920)
-chip_width = 15.36e-3
-chip_height = 9.6e-3
-pixel_size = 8e-6
-bit_depth = 1023
