+++ views/draw_polygon.py
